rename service file
