+++ apps/browser/src/auth/guards/fido2-auth.guard.ts
@@ -9,7 +9,7 @@ import {
 import { AuthService } from "@bitwarden/common/auth/abstractions/auth.service";
 import { AuthenticationStatus } from "@bitwarden/common/auth/enums/authentication-status";
 
-import { PopupHistoryService } from "../../platform/popup/services/browser-router.service";
+import { PopupHistoryService } from "../../platform/popup/services/popup-history.service";
 
 /**
  * This guard verifies the user's authetication status.
