+++ apps/browser/src/platform/popup/services/popup-history.service.ts
